+++ drivers/gpio/gpio_stm32.h
@@ -104,6 +104,20 @@
 #define STM32_PERIPH_GPIOI LL_AHB2_GRP1_PERIPH_GPIOI
 #endif /* CONFIG_SOC_SERIES_.. */
 
+#ifdef CONFIG_SOC_SERIES_STM32F1X
+#define STM32_PINCFG_MODE_OUTPUT	STM32_MODE_OUTPUT | STM32_CNF_GP_OUTPUT | STM32_CNF_PUSH_PULL
+#define STM32_PINCFG_MODE_INPUT		STM32_MODE_INPUT
+#define STM32_PINCFG_PULL_UP		STM32_CNF_IN_PUPD | STM32_PUPD_PULL_UP
+#define STM32_PINCFG_PULL_DOWN		STM32_CNF_IN_PUPD | STM32_PUPD_PULL_DOWN
+#define STM32_PINCFG_FLOATING		STM32_CNF_IN_FLOAT | STM32_PUPD_NO_PULL
+#else
+#define STM32_PINCFG_MODE_OUTPUT	STM32_MODER_OUTPUT_MODE
+#define STM32_PINCFG_MODE_INPUT		STM32_MODER_INPUT_MODE
+#define STM32_PINCFG_PULL_UP		STM32_PUPDR_PULL_UP
+#define STM32_PINCFG_PULL_DOWN		STM32_PUPDR_PULL_DOWN
+#define STM32_PINCFG_FLOATING		STM32_PUPDR_NO_PULL
+#endif /* CONFIG_SOC_SERIES_STM32F1X */
+
 /**
  * @brief configuration of GPIO device
  */
@@ -126,14 +140,39 @@ struct gpio_stm32_data {
 };
 
 /**
- * @brief helper for mapping of GPIO flags to SoC specific config
- *
- * @param flags GPIO encoded flags
- * @param out conf SoC specific pin config
- *
- * @return 0 if flags were mapped to SoC pin config
+ * @brief layout of stm32 gpio registers
  */
-int stm32_gpio_flags_to_conf(int flags, int *conf);
+struct stm32_gpio {
+#ifdef CONFIG_SOC_SERIES_STM32F1X
+	u32_t crl;
+	u32_t crh;
+	u32_t idr;
+	u32_t odr;
+	u32_t bsrr;
+	u32_t brr;
+	u32_t lckr;
+#else
+	u32_t moder;
+	u32_t otyper;
+	u32_t ospeedr;
+	u32_t pupdr;
+	u32_t idr;
+	u32_t odr;
+	u32_t bsrr;
+	u32_t lckr;
+	u32_t afr[2];
+#if defined(CONFIG_SOC_SERIES_STM32F0X) || \
+	defined(CONFIG_SOC_SERIES_STM32F3X) || \
+	defined(CONFIG_SOC_SERIES_STM32F7X) || \
+	defined(CONFIG_SOC_SERIES_STM32L0X) || \
+	defined(CONFIG_SOC_SERIES_STM32L4X)
+	u32_t brr;
+#endif
+#ifdef CONFIG_SOC_SERIES_STM32L4X
+	u32_t ascr;
+#endif
+#endif /* CONFIG_SOC_SERIES_STM32F1X */
+};
 
 /**
  * @brief helper for configuration of GPIO pin
@@ -143,32 +182,6 @@ int stm32_gpio_flags_to_conf(int flags, int *conf);
  * @param func GPIO mode
  * @param altf Alternate function
  */
-int stm32_gpio_configure(u32_t *base_addr, int pin, int func, int altf);
-
-/**
- * @brief helper for setting of GPIO pin output
- *
- * @param base_addr GPIO port base address
- * @param pin IO pin
- * @param value 1, 0
- */
-int stm32_gpio_set(u32_t *base, int pin, int value);
-
-/**
- * @brief helper for reading of GPIO pin value
- *
- * @param base_addr GPIO port base address
- * @param pin IO pin
- * @return pin value
- */
-int stm32_gpio_get(u32_t *base, int pin);
-
-/**
- * @brief enable interrupt source for GPIO pin
- * @param port
- * @param pin
- */
-int stm32_gpio_enable_int(int port, int pin);
+int gpio_stm32_configure(u32_t *base_addr, int pin, int conf, int altf);
 
 #endif /* ZEPHYR_DRIVERS_GPIO_GPIO_STM32_H_ */
-
