+++ soc/arm/st_stm32/stm32f0/CMakeLists.txt
@@ -1,3 +1,2 @@
 zephyr_include_directories(${ZEPHYR_BASE}/drivers)
 zephyr_sources(soc.c)
-zephyr_sources_ifdef(CONFIG_GPIO soc_gpio.c)
