+++ soc/arm/st_stm32/stm32f0/syscfg_registers.h
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef _STM32F0X_GPIO_REGISTERS_H_
-#define _STM32F0X_GPIO_REGISTERS_H_
+#ifndef _STM32_SYSCFG_REGISTERS_H_
+#define _STM32_SYSCFG_REGISTERS_H_
 
 /**
  * @brief
@@ -14,22 +14,10 @@
  *   STM32F030x4/x6/x8/xC,
  *   STM32F070x6/xB advanced ARM ® -based MCUs
  *
- * Chapter 8: General-purpose I/Os (GPIO)
  * Chapter 9: System configuration controller (SYSCFG)
  */
 
-struct stm32f0x_gpio {
-	u32_t moder;
-	u32_t otyper;
-	u32_t ospeedr;
-	u32_t pupdr;
-	u32_t idr;
-	u32_t odr;
-	u32_t bsrr;
-	u32_t lckr;
-	u32_t afr[2];
-	u32_t brr;
-};
+#include "../common/soc_syscfg_common.h"
 
 union syscfg_cfgr1 {
 	u32_t val;
@@ -56,22 +44,14 @@ union syscfg_cfgr1 {
 	} bit;
 };
 
-union syscfg__exticr {
-	u32_t val;
-	struct {
-		u16_t exti;
-		u16_t rsvd__16_31;
-	} bit;
-};
-
-struct stm32f0x_syscfg {
+struct stm32_syscfg {
 	union syscfg_cfgr1 cfgr1;
 	u32_t rsvd;
-	union syscfg__exticr exticr1;
-	union syscfg__exticr exticr2;
-	union syscfg__exticr exticr3;
-	union syscfg__exticr exticr4;
+	union syscfg_exticr exticr1;
+	union syscfg_exticr exticr2;
+	union syscfg_exticr exticr3;
+	union syscfg_exticr exticr4;
 	u32_t cfgr2;
 };
 
-#endif /* _STM32F0X_GPIO_REGISTERS_H_ */
+#endif /* _STM32_SYSCFG_REGISTERS_H_ */
