+++ soc/arm/st_stm32/stm32f1/CMakeLists.txt
@@ -2,4 +2,3 @@ zephyr_include_directories(${ZEPHYR_BASE}/drivers)
 zephyr_sources(
   soc.c
   )
-zephyr_sources_ifdef(CONFIG_GPIO soc_gpio.c)
