+++ soc/arm/st_stm32/stm32f1/soc_registers.h
@@ -8,7 +8,7 @@
 #define _STM32F10X_SOC_REGISTERS_H_
 
 /* include register mapping headers */
-#include "gpio_registers.h"
+#include "afio_registers.h"
 #include "flash_registers.h"
 
 #endif /* _STM32F10X_SOC_REGISTERS_H_ */
