+++ soc/arm/st_stm32/stm32f2/soc_registers.h
@@ -8,6 +8,6 @@
 #define _STM32F2X_SOC_REGISTERS_H_
 
 /* include register mapping headers */
-#include "gpio_registers.h"
+#include "syscfg_registers.h"
 
 #endif /* _STM32F2X_SOC_REGISTERS_H_ */
