+++ soc/arm/st_stm32/stm32f3/soc_registers.h
@@ -9,6 +9,6 @@
 
 /* include register mapping headers */
 #include "flash_registers.h"
-#include "gpio_registers.h"
+#include "syscfg_registers.h"
 
 #endif /* _STM32F3X_SOC_REGISTERS_H_ */
