+++ soc/arm/st_stm32/stm32f4/syscfg_registers.h
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef _STM32F4X_GPIO_REGISTERS_H_
-#define _STM32F4X_GPIO_REGISTERS_H_
+#ifndef _STM32_SYSCFG_REGISTERS_H_
+#define _STM32_SYSCFG_REGISTERS_H_
 
 /**
  * @brief Driver for GPIO of STM32F4X family processor.
@@ -13,33 +13,12 @@
  * Based on reference manual:
  *   RM0368 Reference manual STM32F401xB/C and STM32F401xD/E
  *   advanced ARM(r)-based 32-bit MCUs
- *
- * Chapter 8: General-purpose I/Os (GPIOs)
  */
 
-/* 8.4 GPIO registers - each GPIO port controls 16 pins */
-struct stm32f4x_gpio {
-	u32_t mode;
-	u32_t otype;
-	u32_t ospeed;
-	u32_t pupdr;
-	u32_t idr;
-	u32_t odr;
-	u32_t bsr;
-	u32_t lck;
-	u32_t afr[2];
-};
-
-union syscfg_exticr {
-	u32_t val;
-	struct {
-		u16_t rsvd__16_31;
-		u16_t exti;
-	} bit;
-};
+#include "../common/soc_syscfg_common.h"
 
 /* 7.2 SYSCFG registers */
-struct stm32f4x_syscfg {
+struct stm32_syscfg {
 	u32_t memrmp;
 	u32_t pmc;
 	union syscfg_exticr exticr1;
@@ -49,4 +28,4 @@ struct stm32f4x_syscfg {
 	u32_t cmpcr;
 };
 
-#endif /* _STM32F4X_GPIO_REGISTERS_H_ */
+#endif /* _STM32_SYSCFG_REGISTERS_H_ */
