+++ soc/arm/st_stm32/stm32l0/soc_registers.h
@@ -8,6 +8,7 @@
 #define _STM32L0X_SOC_REGISTERS_H_
 
 /* include register mapping headers */
-#include "gpio_registers.h"
+#include "syscfg_registers.h"
+
 
 #endif /* _STM32L0X_SOC_REGISTERS_H_ */
