+++ soc/arm/st_stm32/stm32l0/syscfg_registers.h
@@ -4,42 +4,21 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef _STM32L0X_GPIO_REGISTERS_H_
-#define _STM32L0X_GPIO_REGISTERS_H_
+#ifndef _STM32_SYSCFG_REGISTERS_H_
+#define _STM32_SYSCFG_REGISTERS_H_
 
 /**
  * @brief
  *
  * Based on reference manual:
  *   STM32L0X advanced ARM ® -based 32-bit MCUs
-
  *
- * Chapter 9: General-purpose I/Os (GPIO)
  * Chapter 10: System configuration controller (SYSCFG)
  */
 
-struct stm32l0x_gpio {
-	u32_t moder;
-	u32_t otyper;
-	u32_t ospeedr;
-	u32_t pupdr;
-	u32_t idr;
-	u32_t odr;
-	u32_t bsrr;
-	u32_t lckr;
-	u32_t afr[2];
-	u32_t brr;
-};
+#include "../common/soc_syscfg_common.h"
 
-union syscfg_exticr {
-	u32_t val;
-	struct {
-		u16_t exti;
-		u16_t rsvd__16_31;
-	} bit;
-};
-
-struct stm32l0x_syscfg {
+struct stm32_syscfg {
 	u32_t cfgr1;
 	u32_t cfgr2;
 	union syscfg_exticr exticr1;
@@ -51,4 +30,4 @@ struct stm32l0x_syscfg {
 	u32_t cfgr3;
 };
 
-#endif /* _STM32L0X_GPIO_REGISTERS_H_ */
+#endif /* _STM32_SYSCFG_REGISTERS_H_ */
